Creating autosleep_hexgrid
This code outputs a Seaborn hexgrid using Autosleep data, scaled to the maximum number of hours in bed. I skipped working with date/time in pandas because it's a bear, and it's far easier to just convert everything to minutes, process it, and convert it back. I've put your file paths back in so you can quickly test if you feel like it. Cell and bin size are set by the number of hours in bed.

Most other categories of time data can be access by copying out the relevant codeblock and just doing a quick replace. Works pretty well.

I'm sure you weren't expecting other additions to this, but hey, got to do something while not sleeping!